Put TodosCalendar into a separate component
