+++ src/components/index.js
@@ -4,3 +4,4 @@ export { TodoDetails } from './TodoDetails/TodoDetails';
 export { TodoItem } from './TodoItem/TodoItem';
 export { CalendarBlock } from './CalendarBlock/CalendarBlock';
 export { TodoBlock } from './TodoBlock/TodoBlock';
+export { TodosCalendar } from './TodosCalendar/TodosCalendar'
